# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface PredictRequest {
  text: string;
}

export interface IcdMatch {
  code: string;
  description: string;
  score: number;
}

export interface PredictResponse {
  results: IcdMatch[];
  ai_available: boolean;
}

export interface SaveRecordRequest {
  input_text: string;
  selected_icd: string;
  icd_description: string;
  confidence_score: number;
}

export interface FhirCoding {
  system: string;
  code: string;
  display: string;
}

export interface FhirCode {
  coding: FhirCoding[];
}

export interface FhirSubject {
  reference: string;
}

export interface FhirCondition {
  resourceType: string;
  subject: FhirSubject;
  code: FhirCode;
  recordedDate: string;
}

export interface Record {
  id: number;
  user_id: number;
  input_text: string;
  selected_icd: string;
  icd_description: string;
  confidence_score: number;
  fhir_json: FhirCondition;
  created_at: string;
}

export interface RecordsResponse {
  records: Record[];
}