# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
  }),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
  name: zod.string(),
});

/**
 * @summary Predict ICD-11 codes from clinical text
 */
export const PredictBody = zod.object({
  text: zod.string(),
});

export const PredictResponse = zod.object({
  results: zod.array(
    zod.object({
      code: zod.string(),
      description: zod.string(),
      score: zod.number(),
    }),
  ),
  ai_available: zod.boolean(),
});

/**
 * @summary Save a confirmed ICD-11 record
 */
export const SaveRecordBody = zod.object({
  input_text: zod.string(),
  selected_icd: zod.string(),
  icd_description: zod.string(),
  confidence_score: zod.number(),
});

/**
 * @summary Get history of saved records
 */
export const GetHistoryResponse = zod.object({
  records: zod.array(
    zod.object({
      id: zod.number(),
      user_id: zod.number(),
      input_text: zod.string(),
      selected_icd: zod.string(),
      icd_description: zod.string(),
      confidence_score: zod.number(),
      fhir_json: zod.object({
        resourceType: zod.string(),
        subject: zod.object({
          reference: zod.string(),
        }),
        code: zod.object({
          coding: zod.array(
            zod.object({
              system: zod.string(),
              code: zod.string(),
              display: zod.string(),
            }),
          ),
        }),
        recordedDate: zod.string(),
      }),
      created_at: zod.string(),
    }),
  ),
});