# lib/api-zod/src/generated/types/fhirCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */
import type { FhirCoding } from "./fhirCoding";

export interface FhirCode {
  coding: FhirCoding[];
}