# lib/api-zod/src/generated/types/fhirCondition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */
import type { FhirCode } from "./fhirCode";
import type { FhirSubject } from "./fhirSubject";

export interface FhirCondition {
  resourceType: string;
  subject: FhirSubject;
  code: FhirCode;
  recordedDate: string;
}