# lib/api-zod/src/generated/types/predictResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */
import type { IcdMatch } from "./icdMatch";

export interface PredictResponse {
  results: IcdMatch[];
  ai_available: boolean;
}