# lib/api-zod/src/generated/types/record.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */
import type { FhirCondition } from "./fhirCondition";

export interface Record {
  id: number;
  user_id: number;
  input_text: string;
  selected_icd: string;
  icd_description: string;
  confidence_score: number;
  fhir_json: FhirCondition;
  created_at: string;
}