# lib/api-zod/src/generated/types/saveRecordRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NAMASTE-ICD11 Clinical Mapping API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveRecordRequest {
  input_text: string;
  selected_icd: string;
  icd_description: string;
  confidence_score: number;
}